+++ controllers/index.js
@@ -1,6 +1,6 @@
 import { searchMovie } from "./searchMovie.js";
 import { poster } from "./poster.js";
-import { letterboxdWatchlist, letterboxdCustomList } from "./letterboxdWatchlist.js";
+import { letterboxdWatchlist, letterboxdCustomList } from "./letterboxdLists.js";
 import { alternativeSearch } from "./alternativeSearch.js";
 import { proxy } from "./proxy.js";
